# api/swagger.yaml
basePath: /
definitions:
...
  model.Entry:
    properties:
      id:
        type: integer
      image:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
  model.Media:
    properties:
      favorite:
        type: integer
      id:
        type: integer
      image:
        type: string
      member:
        type: integer
      popularity:
        type: integer
      rank:
        type: integer
      score:
        type: number
      status:
        type: integer
      title:
        type: string
      type:
        type: integer
      voter:
        type: integer
    type: object
...
  utils.Response:
    properties:
      data:
        type: object
      message:
        type: string
      meta:
        type: object
      status:
        type: integer
    type: object
info:
  contact:
    email: axel.rl.404@gmail.com
    name: Axel
    url: https://github.com/rl404
  description: MyAnimeList database dump and API.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: MAL-DB API
paths:
...
  /search:
    get:
      consumes:
      - application/json
      parameters:
      - description: Entry name
        in: query
        name: query
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Entry'
                  type: array
              type: object
      summary: Quick search
      tags:
      - search
  /search/anime:
    get:
      consumes:
      - application/json
      parameters:
      - description: Anime title
        in: query
        name: title
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Score
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        in: query
        name: score
        type: integer
      - description: Anime type (1=TV, 2=OVA, 3=Movie, 4=Special, 5=ONA, 6=Music)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: type
        type: integer
      - description: Anime airing status (1=airing, 2=finished, 3=upcoming)
        enum:
        - 0
        - 1
        - 2
        - 3
        in: query
        name: status
        type: integer
      - description: Anime rating (1=G, 2=PG, 3=PG13, 4=R17, 5=R, 6=RX)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        in: query
        name: rating
        type: integer
      - description: Anime source (1=original, 2=manga, 3=4-koma, 4=web manga, 5=digital manga, 6=novel, 7=light novel, 8=visual novel, 9=game, 10=card game, 11=book, 12=picture book, 13=radio, 14=music)
        enum:
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        - 11
        - 12
        - 13
        - 14
        in: query
        name: source
        type: integer
      - description: Airing year
        in: query
        name: year
        type: integer
      - description: Season
        enum:
        - winter
        - spring
        - summer
        - fall
        in: query
        name: season
        type: string
      - description: Order (negative means descending)
        enum:
        - member
        - -member
        - title
        - -title
        - score
        - -score
        in: query
        name: order
        type: string
      - description: Start airing year
        in: query
        name: start_year
        type: integer
      - description: End airing year
        in: query
        name: end_year
        type: integer
      - description: Minimum episode count
        in: query
        name: start_episode
        type: integer
      - description: Maximum episode count
        in: query
        name: end_episode
        type: integer
      - description: Minimum duration (in minutes)
        in: query
        name: start_duration
        type: integer
      - description: Maximum duration (in minutes)
        in: query
        name: end_duration
        type: integer
      - description: Producer ID
        in: query
        name: producer
        type: integer
      - description: Genre ID (negative means to exclude the genre)
        in: query
        items:
          type: integer
        name: genre
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Media'
                  type: array
              type: object
      summary: Search anime
      tags:
      - search
  /search/character:
    get:
      consumes:
      - application/json
      parameters:
      - description: Character name
        in: query
        name: name
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Order (negative means descending)
        enum:
        - name
        - -name
        - favorite
        - -favorite
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Entry'
                  type: array
              type: object
      summary: Search character
      tags:
      - search
  /search/manga:
    get:
      consumes:
      - application/json
      parameters:
      - description: Manga title
        in: query
        name: title
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Score
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        - 9
        - 10
        in: query
        name: score
        type: integer
      - description: Manga type (1=manga, 2=light novel, 3=one-shot, 4=doujinshi, 5=manhwa, 6=manhua, 7=oel, 8=novel)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        - 6
        - 7
        - 8
        in: query
        name: type
        type: integer
      - description: Manga publishing status (1=publishing, 2=finished, 3=upcoming, 4=hiatus, 5=discontinued)
        enum:
        - 0
        - 1
        - 2
        - 3
        - 4
        - 5
        in: query
        name: status
        type: integer
      - description: Publishing year
        in: query
        name: year
        type: integer
      - description: Order (negative means descending)
        enum:
        - member
        - -member
        - title
        - -title
        - score
        - -score
        in: query
        name: order
        type: string
      - description: Start publishing year
        in: query
        name: start_year
        type: integer
      - description: End publishing year
        in: query
        name: end_year
        type: integer
      - description: Minimum chapter count
        in: query
        name: start_chapter
        type: integer
      - description: Maximum chapter count
        in: query
        name: end_chapter
        type: integer
      - description: Minimum volume count
        in: query
        name: start_volume
        type: integer
      - description: Maximum volume count
        in: query
        name: end_volume
        type: integer
      - description: Magazine ID
        in: query
        name: magazine
        type: integer
      - description: Genre ID (negative means to exclude the genre)
        in: query
        items:
          type: integer
        name: genre
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Media'
                  type: array
              type: object
      summary: Search manga
      tags:
      - search
  /search/people:
    get:
      consumes:
      - application/json
      parameters:
      - description: People name
        in: query
        name: name
        type: string
      - description: Page
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Order (negative means descending)
        enum:
        - name
        - -name
        - favorite
        - -favorite
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Entry'
                  type: array
              type: object
      summary: Search people
      tags:
      - search
...
schemes:
- http
- https
swagger: "2.0"